Added unit entity to InvoicePosition and updated invoices
The Unit entity has been added to the InvoicePosition entity, allowing users to specify unit for each position in an invoice. This has also impacted the display and generation of invoices, with relevant changes made in views, controllers, and services to accommodate this update. A migration file is also included to update the database schema accordingly.